+++ CHANGELOG.md
@@ -12,6 +12,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 
 - Axes and lines now have `class`/`className` prop.
 - Add `CardinalLine` that use Bézier curve.
+- New polar types: `RadialAxis`, `AngularAxis`, `PolarLine`, and `PolarSector`.
+
+### Changed
+
 - Apply `{...attrs}` props to lines containers.